pyproject.toml
[project]
name = "stock-info"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "feedparser>=6.0.11",
    "firebase-admin>=6.7.0",
    "langchain>=0.1.0",
    "langchain-chroma>=0.1.2",
    "langchain-community>=0.3.21",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.0.6",
    "lxml[html-clean]>=5.3.2",
    "newspaper3k>=0.2.8",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "sentence-transformers>=4.0.2",
    "streamlit>=1.44.1",
    "tweepy>=4.15.0",
]
